Move config under src
