+++ src/yagi.js
@@ -1,5 +1,5 @@
 const Discord = require('discord.js');
-const { token, bisoMixpanel, yagiMixpanel, topggToken } = require('../config/yagi.json');
+const { token, bisoMixpanel, yagiMixpanel, topggToken } = require('./config/yagi.json');
 const yagi = new Discord.Client({
   intents: [Discord.Intents.FLAGS.GUILDS, Discord.Intents.FLAGS.GUILD_MESSAGES],
 });
